basics/04-audio-intelligence/python/requirements.txt
# Speechmatics Batch SDK
speechmatics-batch>=0.4.4

# Utilities
python-dotenv>=1.2.1